chore: refactors detail workout screen
